fix: keep aliases and overlaps as note-only list presets
Aliases and overlaps belong on notes, not folder or notebook readme, so they
are omitted from readme property-key suggestions. When a note already has
either list, do not offer aliases 2 or overlaps 2.
