+++ frontend/tests/utils/noteContentPropertyKeys.spec.ts
@@ -8,8 +8,6 @@ import {
   nextAvailablePropertyKeyForBase,
   nextAvailablePropertyKeyForPreset,
   propertyKeyBaseAndSuffix,
-  richModeKeyDropdownPresetKeys,
-  richModeKeyDropdownPresetKeysForPropertyRows,
 } from "@/utils/noteContentPropertyKeys"
 
 describe("propertyKeyBaseAndSuffix", () => {
@@ -57,11 +55,6 @@ describe("nextAvailablePropertyKeyForPreset", () => {
 
   it("returns the next suffixed key when the base is taken", () => {
     expect(
-      nextAvailablePropertyKeyForPreset("aliases", [
-        propertyRowWithScalar("aliases", "color"),
-      ])
-    ).toBe("aliases 2")
-    expect(
       nextAvailablePropertyKeyForPreset("url", [
         propertyRowWithScalar("url", "https://a"),
       ])
@@ -87,74 +80,6 @@ describe("nextAvailablePropertyKeyForPreset", () => {
     expect(
       nextAvailablePropertyKeyForPreset("url", rows, { excludeRowIndex: 0 })
     ).toBe("url")
-  })
-})
-
-describe("richModeKeyDropdownPresetKeysForPropertyRows", () => {
-  it("matches full list when no rows use preset families", () => {
-    expect(richModeKeyDropdownPresetKeysForPropertyRows(false, [])).toEqual(
-      richModeKeyDropdownPresetKeys(false)
-    )
-    expect(
-      richModeKeyDropdownPresetKeysForPropertyRows(false, [
-        propertyRowWithScalar("status", "ok"),
-      ])
-    ).toEqual(richModeKeyDropdownPresetKeys(false))
-  })
-
-  it("resolves occupied presets to the next suffixed key", () => {
-    const defaults = richModeKeyDropdownPresetKeys(false)
-    expect(
-      richModeKeyDropdownPresetKeysForPropertyRows(false, [
-        propertyRowWithScalar("image", "/a.png"),
-      ])
-    ).toEqual(defaults.map((k) => (k === "image" ? "image 2" : k)))
-    expect(
-      richModeKeyDropdownPresetKeysForPropertyRows(false, [
-        propertyRowWithScalar("wikidataId", "Q1"),
-      ])
-    ).toEqual(defaults.map((k) => (k === "wikidata_id" ? "wikidata_id 2" : k)))
-    expect(
-      richModeKeyDropdownPresetKeysForPropertyRows(false, [
-        propertyRowWithScalar("url", "https://x"),
-      ])
-    ).toEqual(defaults.map((k) => (k === "url" ? "url 2" : k)))
-    expect(
-      richModeKeyDropdownPresetKeysForPropertyRows(false, [
-        propertyRowWithScalar("example of", "[[A]]"),
-        propertyRowWithScalar("example of 2", "[[B]]"),
-      ])
-    ).toEqual(defaults.map((k) => (k === "example of" ? "example of 3" : k)))
-  })
-
-  it("ignores rows with empty keys", () => {
-    expect(
-      richModeKeyDropdownPresetKeysForPropertyRows(false, [
-        propertyRowWithScalar("", "x"),
-        propertyRowWithScalar("  ", "y"),
-      ])
-    ).toEqual(richModeKeyDropdownPresetKeys(false))
-  })
-})
-
-describe("richModeKeyDropdownPresetKeys", () => {
-  it("returns the default rich-mode preset keys", () => {
-    expect(richModeKeyDropdownPresetKeys(false)).toEqual([
-      "aliases",
-      "overlaps",
-      "image",
-      "wikidata_id",
-      "url",
-      "example of",
-      "question_generation_instruction",
-    ])
-  })
-
-  it("appends readme-only keys when isReadmeContext is true", () => {
-    expect(richModeKeyDropdownPresetKeys(true)).toEqual([
-      ...richModeKeyDropdownPresetKeys(false),
-      "title_pattern",
-    ])
   })
 })
